articleEND

+++ TerceraEvaluacion/PHPs/articlesMVC/view/articlesList.php
@@ -11,16 +11,21 @@
          </thead>
          <tbody>
              <?php
-
-                foreach ($article as $value) {
+                if (empty($article)) {
                     echo "<tr>";
-                    echo "<td>" . $value->getId() . "</td>";
-                    echo "<td>" . $value->getName() . "</td>";
-                    echo "<td>" . $value->getDescription() . "</td>";
-                    echo "<td>" . $value->getStock() . "</td>";
-                    echo "<td>" . $value->getPrice() . "</td>";
-                    echo "<td><a href='". BASE_URL ."/articles/delete/". $value->getId() ."' class='btn btn-dark'>Delete</a></td>";
+                    echo "<td colspan='5' class='text text-center'><p class='alert alert-danger'>No articles found</p></td>";
                     echo "</tr>";
+                } else {
+                    foreach ($article as $value) {
+                        echo "<tr>";
+                        echo "<td>" . $value->getId() . "</td>";
+                        echo "<td>" . $value->getName() . "</td>";
+                        echo "<td>" . $value->getDescription() . "</td>";
+                        echo "<td>" . $value->getStock() . "</td>";
+                        echo "<td>" . $value->getPrice() . "</td>";
+                        echo "<td><a href='" . BASE_URL . "/articles/delete/" . $value->getId() . "' class='btn btn-dark'>Delete</a></td>";
+                        echo "</tr>";
+                    }
                 }
                 ?>
          </tbody>
